Added seed in PlanckAquisition
Default value is seed=None: in this case it generates a random seed to
compute Planck noise.
If seed is a number then it is used to generate the noise.

Added seed in PlanckAquisition

Default value is seed=None: in this case it generates a random seed to
compute Planck noise.
If seed is a number then it is used to generate the noise.

Added seed in PlanckAquisition

Default value is seed=None: in this case it generates a random seed to
compute Planck noise.
If seed is a number then it is used to generate the noise.

Added seed in PlanckAquisition

Default value is seed=None: in this case it generates a random seed to
compute Planck noise.
If seed is a number then it is used to generate the noise.