Fix getting the total number of detectors and samplings. (#96)
The 2D cartesian communicator of the acquisition was used in some places
to get the total number of detectors and samplings. The correct
communicators to use are the instrument.comm and sampling.comm
communicators.
Convert the MPI test file to pytest and improve it. The new pytest tests
will live in the tests directory (instead of test).

Co-authored-by: Pierre Chanial <chanial@apc.in2p3.fr>