Bug corrected in Qinstrument.py (somehow lost and found)
Comment by JCH:  The bug was intoduced in 2020 in the commit
f0075d587bd2d3f89063168352eb056ceabdc9f2

In the function _get_response_A() the A field is calculated at the
location requested by the user (in general a TES location) as was
correctly calculated in the first case, where external_A is None.

During the above commit, the field calculation was modified in the
case where external_A is None and was left at all available positions
from the external_A field (which corresponds to the phase and
amplitude of the electric field from each horn in the focal plane. So
instead of having it for a single position we were having it at all
focal plane positions, which resulted in a crash of the code.